src/e/edit/FindAndReplaceAction.java: Let's assume that we can rely on a single interrupt being propagated up the worker thread's stack through the gift of the "reinterrupt itself" idiom from:
http://java.sun.com/j2se/1.5.0/docs/guide/misc/threadPrimitiveDeprecation.html

Waiting for the worker thread to complete wouldn't be any use in preventing the "100 matches" problem.  The only deadlock-free way I could think of waiting for the "finished" callback to complete was to display a hidden dialog which the callback would cancel.  That's over-complicated madness but I failed to come up with anything else, even were I given a free hand to modify or implement an alternative SwingWorker.

In any case, we'd be better off getting on with some new work rather than waiting.  So let's discard any out-of-date results.  There's a reasonably elegant hack we can use to do that.  Sadly, my previously reproducible test case has stopped being reproducible.