src/e/ptextarea/PAnchorSet.java: Elliott reported a linearity violation.  Phil thinks it must be a synchronization problem.  The whole public interface calls checkLinearity, so we can barf there if we're accessed from a thread we shouldn't be.  I didn't want to check this in initially, in case it restricted useful functionality.  One linearity problem doesn't stop a ship.  However, in several hours of using Edit, only one caller caused the exception.
src/e/ptextarea/PTextAreaSpellingChecker.java: Elliott says that:

  in Swing "the various forms of: append, insert, replace, and setText" are thread-safe. ours should be too. reasonable people will assume this. (and let's face, those methods are thread-safe for a good reason.)

So I presume that the highlighting methods aren't, so ours don't have to be.  Perhaps, then, the spelling checker shouldn't be adding highlights from its own thread.

I considered postponing the highlight updates until the end of the spelling check.  That seemed like adding complexity to the normal case for the benefit of the abnormal.  The checkSpelling method gets called lots on incremental updates.  It only gets called once on the whole document, when it's first loaded.  If we perform the spell checker in small stages on the Swing thread, then we avoid inspecting the document, as well as mutating its highlights, without contention protection.  There is no concurrent access to the PTextBuffer in this version, not even to get its length.