src/e/forms/FormDialog.java: revert revision 683 in favor of the null object pattern. Move the addition of the listener to the text fields into the same method as the creation of the listener, for clarity. Remove my "cunning" hack to show that we didn't need to remove anything from the list of text fields in removeTextFieldListeners, which gave the misleading impression that it's acceptable for a collection to be null (instead of empty). Though it's novel, which is usually a bad thing, I like the new code, which treats the list like a queue.
src/e/gui/NoOpAction.java: a suitable null object for ActionListener. Java
tradition would suggest that this be called ActionAdapter, but that's crazy
when there's only one method in the interface, as here. "Dummy" or "No" might
have been alternatives to "NoOp", and "Action" is arguably misleading (though
everyone does this).

src/e/forms/FormPanel.java: default to NoOpAction for the typing timeout.