If you assigned to a global variable in native.make, received an error message and then moved the assignment to java.make, the error message would persist.  This is because make should only regenerate its makefiles when it's read them all, so it has all the rules which may apply to regenerating them.  The $(error) was stopping make in its tracks before it has a chance to regenerate generated/local-variables.make to discover that the error's been fixed.
java.make: Remove generated/local-variables.make if we're about to produce a $(error).  Next time the user runs make, we'll regenerate it and see whether the problem has gone away.  This means that, if you've got a generated/local-variables.make which includes, for example, CFLAGS but you've fixed the underlying problem, then two invocations of make will get you a build.

That would be quite sucky except that, in future, you won't be left with a generated/local-variables.make if it produces an error.  So, in future, you'll only need a single invocation of make to notice the fix.

java.make: Make the error message for an invalid local variable access more informative.  We certainly don't seem to be able to rely on the line numbers which make outputs.

native.make: The purpose of this file has become a little clearer.