src/e/ptextarea/PIndenter.java: Treat the " * " in the middle of doc-comments as indentation.  That will let us trim trailing space off the line in the indentation fixer without removing this necessary space.  It's also the first step towards improving our backspace-across-line-ending behavior.  If the user hits backspace and we know we're at the start of the indentation, then we can take them straight back to the end of the previous line.
We shouldn't be recognizing the " * " pattern unless shouldContinueDocComments() is true.  Which is to say that this behavior wants to move, at least partly, to PSimpleIndenter.  Also, someone will hopefully suggest how I should have coped with the allegedly impossible situation where I currently throw an IllegalArgumentException.

src/e/ptextarea/PIndenter.java: Trim trailing space when fixing the indentation.  Change the algorithm for preserving the selection such that offsets within the indentation move to the end of the replacement indentation.  That feels much better when you hit enter in the middle of the indent.  Instead of effectively preserving the horizontal position, we move the caret to where you'd want to start typing.

fixIndentationAt has become too complicated because I've tried to fix the indentation and trim the trailing space in one operation.  I was concerned that nesting compound edits might not work.  Some of our callers already create a compound edit.  I think I should have just done this as two separate edits and worried about compounding later.

src/e/ptextarea/PNewlineInserter.java: If we're saying it's a reasonable thing to hit enter in the middle of an indent and expect Edit to do the right thing, then we need to fix the indentation of the line before the newline as well as the line after.

src/e/ptextarea/PSimpleIndenter.java: The indentation of the mid-comment lines is already taken care of, so this code gets a little simpler.  The first change here is a fix for an unrelated bug which meant that the line following a empty mid-comment line wouldn't be regarded as being in the comment.

In summary, this needs tidying up when I'm less tired and after someone's commented about the nested compound edit issue, but I believe this fixes both inserting-in-the-indent infelicities that Elliott reported.