src/terminator/LogWriter.java: Lose the colons from the date format in the log file names.  I was complaining just this evening that it's difficult to disentangle the line number which grep appends from the rest of the filename and these colons weren't helping.  No-one could ever find anything in the .terminal-logs directory without using grep anyhow.
This lets terminator run on cygwin.  Well, it would if cygwin worked for the JNI DLL.  My best guess from what I've read of the source and on the interweb that there's a bug in the cygwin DLL initialization when it's loaded into a non-cygwin process like the JVM.  I think it's designed to cope.