universal.make: In an unsuccessful effort to provoke the strerror_r warning from the linker, I've reinstated -pedantic but only for C++.  That was just my immediate motivation: really I'm doing this on the general principle that we should compile with maximum warnings and pedantry unless there's a compelling reason otherwise.  In particular, I'm sure that -pedantic found a number of things which needed tidying up in the BlueArc code.  (I've found the relevant check-in comment but, sadly, I wasn't specific:
  Everything seems to compile with -pedantic now.  NFS isn't pretty.

That probably means that a lot of fixes were needed.)  Elliott had previously removed -pedantic on the grounds that it produces a bogus warning for Objective-C and that the manual doesn't clearly specify the purpose of -pedantic.  It doesn't doesn't make it any less distressing but it doesn't clearly specify the purpose of "volatile" either.

(I need to leave the stutter in the last paragraph to illustrate a possible PTextArea painting problem.)