src/e/forms/FormDialog.java: src/e/forms/FormPanel.java: src/e/gui/AlphabeticComboBoxModel.java: src/e/gui/ChronologicalComboBoxModel.java: src/e/gui/EHistoryComboBoxModel.java: src/e/gui/EPopupMenu.java: src/e/gui/FilteredListModel.java: src/e/gui/JTextComponentSpellingChecker.java: src/e/gui/SortedComboBoxModel.java: src/e/gui/WindowMenu.java: src/e/ptextarea/PHyperlinkTextStyler.java: src/e/ptextarea/PLineList.java: src/e/ptextarea/PTextArea.java: src/e/ptextarea/PTextAreaSpellingChecker.java: src/e/ptextarea/PTextBuffer.java: src/e/ptextarea/PTextWindow.java: src/e/tools/JarExplorer.java: src/e/toys/world/Gazetteer.java: src/e/util/Bag.java: src/e/util/FileSearcher.java: src/e/util/Launcher.java: src/e/util/NumberDecoder.java: src/e/util/ProcessUtilities.java: src/e/util/SpellingChecker.java: src/e/util/StringHistory.java: src/e/util/StringUtilities.java: use generics.
src/e/ptextarea/PAnchorSet.java: use generics. I think -Xlint:unchecked turned
up an error here, in that we weren't using the (broken) AnchorComparator for
our binary search.

src/e/ptextarea/PCLikeTextStyler.java: fixing this was a bit awkward, and
involved the *addition* of a couple of casts, because of the confusion between
PLineSegment and PTextSegment which seems to extend into the code, not just
the names.

src/e/util/InstanceTracker.java: I haven't completely fixed this; I'm not sure
I can.