parent.make: I've had something like this file in the parent directory of all my software.jessies.org projects for a few days.  I just came to update it and thought "maybe this'd be useful to someone else" shortly followed by "it's annoying that this isn't version controlled, now it's more than three lines long".  Then I realized I *can* have it in salma-hayek if ../Makefile just contains:
  include salma-hayek/parent.make
  
The change I made today was to make the update rule wait for completion of its background jobs, using the trick I came up with for reachable_build_servers, of piping all their output to the same place.  That let me change the default target to update as well as build.

The initial check-in comment would have said something like "a Makefile to update all the Subversion repositories in parallel, so halving (at a guess) the total time it takes me to update-and-build".  It current supports targets build, update, clean and the default target.  It might happen to support other targets too, like bindist.