src/e/ptextarea/PSimpleIndenter.java: Fix my preceding "fix for an unrelated bug".  It was considering any line whose active part started with an asterisk as being in a doc-comment, and so erroneously matching eg "*pointer = &object;" and prepending "* " to following lines.  I think I must have been trying to fix the behavior on pressing Enter at the end of an empty doc-comment line *whose trailing space had been removed*.  We can do that by seeing if the previousLine just consists of an asterisk.
That doesn't work perfectly because the asterisk isn't recognized by the indentation fixer as being part of the indentation.  So, when its indentation is fixed, the line is given an extra "* ".  That's more benign that either of the previous behaviors.  (The behavior before the preceding change was to terminate the doc-comment.)