src/e/ptextarea/RegularExpressionStyleApplicator.java: make it possible for subclasses to decide exactly what kind of PTextSegment is created.
src/e/ptextarea/PStyle.java: add a new style for unprintable characters.

src/e/ptextarea/UnprintableCharacterStyleApplicator.java: a new style applicator that recognizes runs of unprintable characters and creates a new kind of text segment for them.

src/e/ptextarea/UnprintableCharacterTextSegment.java: a new kind of text segment for runs of unprintable characters. This works by overriding getText to return not the unprintable characters themselves, but an escaped version of them. This almost works perfectly, but there's code elsewhere that measures the widths of the characters in the text buffer, rather than asking the text segments about the widths. So if you try to select an unprintable character, which is now six characters wide, as you sweep successive characters in the visible representation of the unprintable character, you'll see characters past the unprintable start to get selected. This is perhaps what Phil was alluding to when he suggested at some point that that code needed rewriting and moving out of PTextArea.

src/e/ptextarea/PTextArea.java: add a style applicator for unprintable characters by default. The selection behavior needs work, but this is still much more useful than these characters simply being invisible.