src/e/ptextarea/PTabSegment.java: Fail noisily rather than silently doing the wrong thing when given a zero length tab segment.  I've scratched my head for a while without coming up with a tasteful way to fix getDisplayWidth.  I contemplated banning zero-length segments more widely but, given that getCharOffset manages to be tasteful (though it could do with a better name for "startX"), that seems like the tail wagging the dog.  It would also go against the grain to ban zero, the most elegant number.
src/e/ptextarea/PTextArea.java: Stop scanning over segments when we find the offset we're looking for at the "end" of one, rather than calculating the display width of a zero-length subsegment of the next segment.  I don't see any difference in elegance and this way should be marginally more efficient, though the motivation is to avoid the bug in PTabSegment noted above.

This fixes the appearance of the selection of tabs which don't start on tab-width boundaries.  Before the fix, the selection appeared to extend to the previous tab-width boundary.