universal.make: install-everything.sh failed again this morning on tencents due to it "only" having the latest Debian make.  (I don't understand why no-one else has reported this problem, which has happened before on tencents.  It's definitely down to the make version.)  These are the error messages:
make: Entering directory `/usr/local/www.jessies.org/terminator'
../salma-hayek/universal.make:431: *** makefile bug: local variable HEADERS from scope pty (with value "/usr/local/www.jessies.org/terminator/native/all/pty/PtyGenerator.h /usr/local/www.jessies.org/terminator/native/all/pty/errnoToString.h") was referred to in scope pty.  Stop.
make: Leaving directory `/usr/local/www.jessies.org/terminator'
building terminator

The thought occurred that the dist we're downloading is supposed to include the generated files for the architecture it was built on.  Looking at the terminator/Makefile, that adds "pty" to BINDIST_FILES but that directory hasn't been used for some months now.  That led me to investigate the behavior of BINDIST_FILES and to conclude that it doesn't work and it's been broken for ages.  So no-one probably cares what it does, so if I make it work by making it do something simpler but reasonable, there probably won't be too much shouting.  So:

(1) I'm removed the option to customize BINDIST_FILES in the individual makefiles.  Strictly speaking, it's still there but now undocumented as nothing now needs it (I think).  I'll check in versions of scm/Makefile and terminator/Makefile which remove the setting of this variable in a minute.

(2) Use $(shell) rather than backquoting to get the date.  This means that make knows what tar file it's building.

(3) That lets me separate the rule for building the tar file from the rule from uploading the dist, which lets me test the rule for building the tar file without waiting for ssh all the time.

(4) LdLibraryPathPollution.class demonstrated a flaw in our usage of the jar command.  -C only seems to apply to the next argument, not to the rest of the arguments.  We're using cd and -C, which seems like madness.  Just use cd, which has a more widely understood and less surprising semantics.

(5) Add the doc directory, if it exists, to the bindist, obsoleting part of the terminator/Makefile.

(6) Add the bin directory, on the assumption that it exists in all our projects, to the bindist, obsoleting another part of terminator/Makefile and all of scm/Makefile.  This adds everything in the bin directory rather than just the named files.  Presumably the other scripts in the bin directory are useful.

(7) Now we have universal.make, we have a complete list of the native targets we've built.  Add those to the bindist.  This should add just the shared libraries and executables.  With no need to find(1) - neat.

(8) find $(PROJECT_NAME).jar was causing an error.  Stop trying to be clever about treating BINDIST_FILES as a list of files and directories.

(9) We use Subversion now, not CVS, so make bindist should exclude .svn directories too.

(10) Add the generated directory to make dist - to hopefully fix the original problem.  I'm not sure if everything in there really wants to be added.  I've made a new dist and verified that this *doesn't* fix the problem.  That doesn't surprise me: I expected the problem was in make's first pass rather than building the targets which are out-of-date.

Oh well, these changes are still good, just insufficient.