parent.make: There's no uploading yet but now "make dist" in the parent directory builds one big .tar.gz file with what would be in all of the projects' .tar.gz dist files, having cleaned and built before to ensure that we don't get any spurious content.  It also cleans and builds afterwards to get rid of the ChangeLog files which clutter checkintool.
I've added some of the make boilerplate and split variables and rules in the normal manner.  I've always switched to more explicitly request recursion when desired.  This was all because I noticed that the order-only dependencies aren't working and I was trying to put things in more normal form to see if that fixed it.  It didn't and I don't know why this is broken.  Maybe order-only prerequisites don't work for .PHONY targets.  I need to report that, if so.  At the moment everything seems to work OK regardless because make tends to build the prerequisites in order unless it has reason to do otherwise.