native/all/pty/terminator_terminal_PtyProcess.cpp: The "let them set fields" check-in didn't explain why:
  I think an important step in the direction of a good JNI idiom is that native methods should be void. They can set fields if they like.

The justification, as I understood it, was that a bogus return value would otherwise have to be invented when an exception is thrown.  I don't find this is a very compelling justification.  Sufficiently compelling, yes, but I can imagine someone coming up with a more compelling reason why a JNI function should return a value.