universal.make: I noticed a spurious := being (correctly) preserved in the preceding check-in.  Correctly preserved but spurious nonetheless, so out it comes, along with the other one I found in this file.
There's a rather complicated and entirely unwritten rule which determines when to use :=.  A check-in comment probably isn't the right place to record it.  Still, basically the rule is to use = unless there's a good reason otherwise.  Doing something external to make is the most usual good reason otherwise.  The two most common examples of that are $(wildcard) and $(shell).  The main reason for using := with those is efficiency but it also saves thinking about non-idempotence.  Yeah, yeah, efficiency but it gets significant on big projects (BlueArc-scale, not salma-hayek scale).  There are quite a few other good reasons for using :=.  Generally they warrant a comment.

Why should = be the default assumption?  The lazy evaluation which results can be more efficient.  make doesn't have any structure syntax or any scoping, so things get disorganized and out-of-order.  So, in part, = is a coping strategy for avoiding dependency on correct ordering.