universal.make: It looks to me like we've been producing an error message during install-everything.sh on machines with svn installed for some time:
svn: '.' is not a working copy

We were doing the offending svn info operation simply to cope with the Subversion "trunk" nonsense, that we're no longer using.  Remove support.

Now the list of files which go to make up a "make dist" are mentioned in a dependency (since I split up "make dist"), the list is always evaluated, even in the make invoked by install-everything.sh, which can't talk to the revision control system.