native/all/pty/terminator_terminal_PtyProcess.cpp: We've lost toString to an eponymous header.  Collapse the " failed" + errnoToString() duplication into...
native/all/pty/UnixException.h: ... a class which just does that.  Elliott suggested "unix_error" for the name.  I didn't like "error" because this is an exception class and it feels to me like it should say what it is on the tin.  I didn't like the lowercase because that made it look like it was pretending to be part of the standard library.

native/all/pty/PtyGenerator.h: We can use UnixException here too, now it's abstracted.  That will give us more consistent albeit less literate error messages.  Especially given that these error messages won't ever happen and, if they do, they're not going to mean much without the source, I think consistency is more useful.