native/all/pty/UnixException.h: A missing public is one of the consequences I get for trying to program outside my normal idiom.  "message" is a better name than "str".
native/all/pty/PtyGenerator.h: What's good for the parent would be good for the child.  Report the errno-setting errors in runChild via UnixException.  This reduces the number of calls to exit() to one, which I've commented to appease my worry that we'd quit Edit if that line were ever executed.  (Perhaps, one day, Edit will start terminator windows from within its VM.)  That led naturally to removing the return value from runChild, a detail which previously inspired two comments in the source.

The error messages are now all (I think) consistent, except for the pseudo-terminal ones, where we replace the file descriptor with the ptsname, and the execvp one where we output a more friendly message.  The execvp one is the only one of these exceptions that I think anyone's likely to see.

clientPanic was an odd name.  It begins with "c" and has the same number of letters as "child".  I wonder if "childPanic" was the intention.

I'm not convinced by this toString malarky but it's certainly concise.  Any inefficiency of string copying doesn't matter in exception-throwing code.  There's no must-not-malloc constraint here either: we're not writing device drivers.  Hmmm... maybe it'll work after all.