post-commit: Duh, I wanted the path of the repository, not its name.  Perhaps renaming "repo" to something more specific will help me avoid that mistake in future.
When I hacked the checked-out copy of the script to echo the command it would execute and ran it with:

  mad@jessies:~$ /home/software/checked-out/salma-hayek/post-commit /home/software/svnroot/salma-hayek 888

I saw:

  find checked-out/db -name log.* -not -perm -g+w

That suggests that repository_name somehow changed to be "checked_out".  I wonder if that's due to the script updating itself (or whether it's some dynamism in Ruby that I haven't understood).