src/e/edit/EditServer.java: add a new "open-and-block" verb that waits for the window to be closed before closing the connection.
bin/edit: 
bin/edit-and-block: add a "--block" option to the "edit" script, and add a new "edit-and-block" script for use with the $EDITOR environment variable. This lets you use Edit to edit your crontab. (There may be other applications, but I don't have any. "bk resolve", though visually confusing, waits for you to give it the "C" command anyway. And I haven't used "cvs commit", the other place I've wanted this functionality, since a couple of hours in to writing SCM.)