src/e/ptextarea/PUnderlinedTextSegment.java: a complaint that links to errors in Edit's errors window weren't working turned out to be the final provocation I needed to remove PUnderlinedTextSegment.
src/e/ptextarea/PStyle.java: isUnderlined becomes a property of the individual styles (currently true only of the hyperlink style).

src/e/ptextarea/RegularExpressionStyleApplicator.java: we no longer have to construct hyperlink text segments as a special case.

src/e/ptextarea/PTextSegment.java: here we gain the code to draw the underline. We also need to preserve the linkAction and toolTip when subSegment is called (this was why our errors window links weren't doing anything), and it's a good idea if toString shows the value of our linkAction field too. Note that I've deliberately left the linkAction field uninitialized (rather than setting it to NoOpAction.INSTANCE) because it's an error for someone to invoke linkClicked if we're not actually a link, and the NullPointerException was the only external evidence of the error in PUnderlinedTextSegment.subSegment here.