src/e/ptextarea/PLineEvent.java: Add a 'completely replaced' type of PLineEvent, so the PTextArea can receive such events when setText has been called.
src/e/ptextarea/PLineList.java: Fire the new type of event when the text's been completely replaced.

src/e/ptextarea/PLineListener.java: Add a new listener method for same.

src/e/ptextarea/PTextArea.java: Add a clearSegmentCache method, and call it when we know the buffer's been replaced.  This solves the problem of the buffer temporarily being drawn in a corrupted state on 'reload from file'.  Also rename 'flushSegmentCacheFrom' to 'clearSegmentCacheFrom' in the name of accuracy and consistency.