src/terminator/terminal/PtyProcess.java: The "throws IOException" specifiers were put in when the native code was throwing IOException.  Now the generated proxy is throwing RuntimeException instead.  It's useful to fix the specification because RuntimeException isn't checked, unlike IOException, which requires catching everywhere.
This change begs the question of whether any of the code which is explicitly catching IOException should be generalized.  All the ones in TerminalControl look OK to me.  There are two in JTerminalPane which are currently mandated by other throw specifiers but which might warrant a closer look.

src/terminator/terminal/TerminalControl.java: Wait for the ptyProcess from the same thread which spawned it, making it so that normally exited bash shells have their pane closed on Linux 2.4.  Presumably this also broke when Elliott tidied up the handling of WIFSIGNALED and the like.  The conditional just below this change suggests that he must have fixed things such that we were correctly noticing a normal exit for the first time.  This seems to have exposed a couple of latent issues.  (The bug in Linux worked-around here and the bug in this function fixed by the previous change.)

Does this expose us to a potential deadlock?  Only if the Swing thread is waiting on the Terminal connection listener, which I don't think it does.  I've tested this change with various exit commands on Linux 2.6 and 2.4.  It seemed preferable to the symmetric alternative: spawning the ptyProcess on the Terminal connection listener thread.  A naive implementation of that exposed various variables to access before initialization.

Another alternative would be to disclaim support for Linux 2.4.  If this change causes trouble, that may well be what we'll do.  2.6.11 has been good to me.