native/all/pty/terminator_terminal_PtyProcess.cpp: Remove obsolete Win32 stubs.  I've worked out how to safely load the cygwin-using version of the pty shared library (with cygwin-1.5.18-1) but it's not useful for terminator's purposes.
For one thing, cygwin's fork runs the original executable again which is expected to notice that it's been run as a fork.  java.exe not being linked with cygwin, it doesn't notice, and just produces usage information and quits.  The interweb leads me to believe that you can call Java from a C program with JNI.  That suggests a possible way round this problem using a cygwin-linked launcher instead of java.exe.

Then there's the matter of what sort of file descriptors JNI's wanting.  A goodly few minutes searching on google didn't really leave me any the wiser as to whether it's after Win32 HANDLEs or MSVCRT file descriptors or what.

The cygload example, the cygwin FAQ and how-cygtls-works.txt explain the problem with naively loading a cygwin-using DLL.  Well, actually, they explain the problem with naively loading the cygwin1.dll itself.  The interweb doesn't show much (any?) interest in loading a cygwin-using DLL.  I got around the problem of cygwin overwriting the bottom of the calling stack by creating a shim-pty.dll, compiled with mingw32, which creates a thread with a 4KiB array on the stack then calls Win32's LoadLibrary("pty.dll").  That then lets the Java System.loadLibrary("pty") bind to the DLL without overwriting one of Java's stacks.  (I called Thread.sleep between loading the DLLs to wait for this thread to do its business.)

If you want to try a pathological example program like this, beware of the trap I ran into yesterday.  cygwin1.dll has a shared memory area which it uses to remember how to subvert new threads into calling its own entry point first.  This area is initialized the first time cygwin1.dll is loaded.  The initialization requires waiting for a thread to complete.  If cygwin1.dll is being loaded out of DLL_PROCESS_ATTACH or as a DLL dependency of another DLL, there's a kernel mutex which prevents threading.  So startup deadlocks.  The easy answer is to run your pathological example from bash.