src/terminator/terminal/PtyProcess.java: Re-implement simple versions of FileInputStream and FileOutputStream in my pidgin Java, deferring to JNI methods.  Use these PtyI/OStreams if the JNI code declines to return us a FileDescriptor for use, as before, with FileI/OStreams.
native/all/pty/terminator_terminal_PtyProcess.cpp: Implement these in terms of cygwin's read and write syscalls.  These implement a substantial fraction of the pseudo-terminal behavior on cygwin, so we have to call the cygwin versions of these calls rather than using the underlying Win32 handles.

This lets me check-in code which (hurrah!) works on cygwin without changing the way the code works on proper operating systems.  We could do away with the FileI/OStreams and the FileDescriptor hack.  The cygwin-compatible code does work as far as I tested it on Linux.  Simply change the (second) #ifdef __CYGWIN__ to #if 1 to try it.

PtyInputStream and PtyOutputStream feel like they ought to be non-nested classes (to me).  The only reason I didn't try to do it like that is that I didn't want to have to do any tinkering with the makerules to get multiple JNI classes working in the same DLL.  Maybe it'd just work, I don't know, I didn't even try.

The one key ingredient that remains unchecked-in is the cygwin-linked "java.exe" launcher.