simonj wanted middle button paste and implicit copying of the selection to the clipboard on Win32.  After using terminator on cygwin for a while in earnest, I see that that's what my subconscious is also expecting.
I couldn't just hack the existing methods in good conscience.  The existing methods - copyToSystemSelection and pasteSystemSelection - did one thing and that thing was exactly what the name suggested.  Even when on a platform where you exclaim "I have no SystemSelection!", the methods do the obvious thing - nothing.

src/terminator/view/highlight/SelectionHighlighter.java: I think what we really want, then, is a new method which is called when the selection is changed to copy it somewhere appropriate.

src/terminator/view/JTextBuffer.java: And a method which is called when the middle button is pressed (if one has a middle button) to paste whatever's appropriate, if anything.

src/terminator/view/JTerminalPane.java: I've tidied up the related names too because it's confusing to have methods called plain "copy" and "paste" when there are related methods in the same class which have "copy" in the name.  I think when I introduced those names, I was trying to keep a change textually small.  That's not necessarily a good indicator of simplicity.

src/terminator/TerminatorMenuBar.java: What's this commented-out pasting code for?  Why am I tracking an API change in code which looks like a TODO?