src/terminator/view/JTerminalPane.java: Fix backspace under Windows.  This makes (cygwin) vim and (native) jdb work.
Looking at the rxvt-2.7.8 source and trying to ignore the complexity, it basically always seems to send ^?.  (If you press Shift or Ctrl, you get ^H.)  It does look at the tcgetattr but, I think, only to copy the settings from the initial tty to the spawned ones.  The user's free to change the stty erase setting later so, unless we re-read it all the time, we won't know what the latest setting is.  It seems wrong to second-guess the user like that anyway.  The stty settings are the user's mechanism for coping with problems between the terminal and the equipment on the other end.  If we start changing the ground under their feet, that'll make things more confusing than they already are.

The Linux console and X, Mac OS X and Solaris all seem to use ^? by default.  I didn't know before today that ssh transfers the key mappings (search for ssh "terminal modes").  Telnet doesn't.