src/terminator/view/JTextBuffer.java: Lose the 1024 character hack, which just made the problem less likely, probably by slowing us down.  Lose the thread to sendUtf8String, where it will catch all writes, not just big ones.
src/terminator/terminal/TerminalControl.java: Always write to the pipe on a separate thread, so pipe waits can't block the event dispatch thread.  With this, I've failed to cause a lock-up using big pastes and concurrent keyboard input in a couple of minutes of trying.  I think this is the proper fix for that problem.