www/index.html: Replace references to Unix with POSIX or remove, as appropriate.  Replace the reference to C with one to C++.  Stick to the point in the portability bullet.  Replace "Mac OS" with "Mac OS X", somewhat anally, so no-one thinks we're into retro portability.  Be honest about how much effort we've spent on ensuring that Terminator is easy to run.
Mention Windows.  We don't have a very good out-of-the-box experience on Windows but that would be fixable.  You never know: if we mention it, we might even get someone volunteering to help.

I considered mentioning TeraTerm and SecureCRT but I'm only really familiar enough with PuTTY to say that I think we're a good replacement.  To be clear, I don't have any reason for thinking that it'd be worth even trying the other two programs I've seen people use.  Our QA guys once told me that they use TeraTerm because it offers better scripting capabilities than PuTTY, which stores configuration in the Windows Registry.  I *suspect* that terminator already offers just as good scripting.

Mention UTF-8.  Who knows if we really support it properly?  It's daft not to advertise it when it was only ten minutes work to implement something that's at least partly working.  *Billions* of people are doubtless waiting for this in order to switch.