native/all/launcher/launcher.cpp: A free Java launcher, written only using the information in the JNI web pages.  This compiles and works on Linux and Cygwin and compiles on Mac OS X but doesn't work.  The terminator/bin/terminator modification I'm about to check-in only uses the executable on Cygwin.  There's been talk of rolling our own cross-platform launcher, so native/all seemed like the right place.  (Compiling on Linux produced an extra warning which didn't occur on Cygwin and vice versa.)
This is much dumber and simpler than the Sun launcher.  It hardcodes the choice of JVM for one thing.  It doesn't understand -cp or terminator.Terminator for another.  Instead, it delegates directly to the JVM, which does understand the alternative forms -Djava.class.path and terminator/Terminator.

I did discover a couple of non-obvious things, which I've commented on in the source.  Of particular interest should be the ExceptionOccurred business.  That probably indicates a change we need to make in the terminator JNI or, more probably, in the javahpp-generated infrastructure.

universal.make: On Linux, dlopen is provided by -ldl.  On Cygwin and Mac OS X, it seems to be provided by default.