paulb said today "that's what I don't like about terminator" and it turned out that he meant that delete seems to work synonymously with backspace, as far as bash on Linux is concerned (at least).
src/terminator/view/JTerminalPane.java: Our terminfo entry says we return this string for the delete key.  This doesn't fix the problem but it's got to be a step in the right direction.  Though I have some other hacks make it difficult to be sure, delete now does backspace *and* what you'd expect from delete.  Odd!