bin/install-everything.sh: Previously, we've always been removing /root/.terminal-logs/.terminator-server-port which does something of no use to man nor beast.  I could have tried to find out who's running sudo install-everything.sh but that wouldn't work for the cronjob I run on johnh's machine.  I could always move that to his crontab but the easiest thing seemed to find out all the users who are homed on this machine's local disks and make sure they automatically get a new terminator in the morning.  This should finally stop John using a month-old version.  That's working fine but means he rarely has the latest fixes.
I've no qualms about implementing part of this script with Perl.  One day it'll be translated into Ruby and that's more straightforward from (this sort of simple) Perl than bash.

Clean the project before building it so that we'll rebuild any binaries from the distribution.  libc seems to be changing quite a bit in Debian wobbly-foal at the moment (as is apparently their wont after a new release) and each change is incompatible with the last, causing this sort of error:

/usr/lib/libstdc++.so.6: version `GLIBCXX_3.4.4' not found (required by /usr/local/www.jessies.org/terminator/generated/native/all/pty/Linux/libpty.so)

All the machines on which we run install-everything.sh should be capable of building all the projects.